+++ composites-fiberglass/src/main/kotlin/ru/ldralighieri/composites/fiberglass/grid/horizontal/FiberglassLazyHorizontalGrid.kt
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2023 Vladimir Raupov
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ru.ldralighieri.composites.fiberglass.grid.horizontal
+
+import androidx.compose.foundation.gestures.FlingBehavior
+import androidx.compose.foundation.gestures.ScrollableDefaults
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyGridState
+import androidx.compose.foundation.lazy.grid.LazyHorizontalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.foundation.lazy.grid.rememberLazyGridState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import ru.ldralighieri.composites.fiberglass.model.FiberglassItem
+import ru.ldralighieri.composites.fiberglass.model.FiberglassLazyGridItemSlots
+
+/**
+ * Fiberglass lazy horizontal grid Composite
+ *
+ * @param items List of [FiberglassItem] items.
+ * @param itemSlots FiberglassLazyHorizontalGrid [slots map][FiberglassLazyGridItemSlots].
+ * @param modifier The modifier to apply to this layout.
+ * @param state The state object to be used to control and observe the grid state.
+ * @param contentPadding A padding around the whole content.
+ * @param reverseLayout Reverse the direction of scrolling and layout.
+ * @param horizontalArrangement The horizontal arrangement of the layout's children.
+ * @param verticalArrangement The vertical arrangement of the layout's children.
+ * @param flingBehavior Logic describing fling behavior.
+ * @param userScrollEnabled Whether the scrolling via the user gestures or accessibility actions
+ * is allowed. You can still scroll programmatically using the state even when it is disabled.
+ */
+@Composable
+fun FiberglassLazyHorizontalGrid(
+    items: List<FiberglassItem>,
+    itemSlots: FiberglassLazyGridItemSlots,
+    rows: GridCells,
+    modifier: Modifier = Modifier,
+    state: LazyGridState = rememberLazyGridState(),
+    contentPadding: PaddingValues = PaddingValues(0.dp),
+    reverseLayout: Boolean = false,
+    horizontalArrangement: Arrangement.Horizontal =
+        if (!reverseLayout) Arrangement.Start else Arrangement.End,
+    verticalArrangement: Arrangement.Vertical = Arrangement.Top,
+    flingBehavior: FlingBehavior = ScrollableDefaults.flingBehavior(),
+    userScrollEnabled: Boolean = true
+) {
+    LazyHorizontalGrid(
+        rows = rows,
+        modifier = modifier,
+        state = state,
+        contentPadding = contentPadding,
+        horizontalArrangement = horizontalArrangement,
+        verticalArrangement = verticalArrangement,
+        flingBehavior = flingBehavior,
+        userScrollEnabled = userScrollEnabled
+    ) {
+        items(
+            items = items,
+            key = { it.id },
+            contentType = { it::class.simpleName }
+        ) { item ->
+            itemSlots[item::class]?.let { it(item) }
+        }
+    }
+}