+++ composites-fiberglass/src/main/kotlin/ru/ldralighieri/composites/fiberglass/model/FiberglassAliases.kt
@@ -16,9 +16,12 @@
 
 package ru.ldralighieri.composites.fiberglass.model
 
+import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.lazy.LazyItemScope
+import androidx.compose.foundation.lazy.grid.LazyGridItemScope
+import androidx.compose.foundation.lazy.staggeredgrid.LazyStaggeredGridItemScope
 import androidx.compose.runtime.Composable
 import kotlin.reflect.KClass
 
@@ -57,3 +60,28 @@ typealias FiberglassRowItemSlot = @Composable RowScope.(item: FiberglassItem) ->
  * FiberglassRow slots map alias
  */
 typealias FiberglassRowItemSlots = Map<KClass<out FiberglassItem>, FiberglassRowItemSlot>
+
+/**
+ * Fiberglass[Vertical|Horizontal]Grid slot alias
+ */
+typealias FiberglassLazyGridItemSlot = @Composable LazyGridItemScope.(item: FiberglassItem) -> Unit
+
+/**
+ * Fiberglass[Vertical|Horizontal]Grid slots map alias
+ */
+typealias FiberglassLazyGridItemSlots = Map<KClass<out FiberglassItem>, FiberglassLazyGridItemSlot>
+
+/**
+ * Fiberglass[Vertical|Horizontal]Grid slot alias
+ */
+@OptIn(ExperimentalFoundationApi::class)
+typealias FiberglassLazyStaggeredGridItemSlot =
+    @Composable LazyStaggeredGridItemScope.(item: FiberglassItem) -> Unit
+
+/**
+ * Fiberglass[Vertical|Horizontal]Grid slots map alias
+ */
+@OptIn(ExperimentalFoundationApi::class)
+typealias FiberglassLazyStaggeredGridItemSlots =
+    Map<KClass<out FiberglassItem>, FiberglassLazyStaggeredGridItemSlot>
+